Better webpack config
