+++ .gitignore
@@ -8,3 +8,4 @@ components/
 website/.bundle/
 website/vendor/
 build/*.test.*
+build/*.debug.*
