+++ config/karma.conf.js
@@ -4,7 +4,7 @@ module.exports = function(config) {
   config.set({
 
     // base path, that will be used to resolve files and exclude
-    basePath: 'spec/',
+    basePath: '../spec/',
 
     // frameworks to use
     frameworks: ['jasmine'],
@@ -61,6 +61,12 @@ module.exports = function(config) {
     },
 
     webpack: {
+      module: {
+        preLoaders: [{
+          test: /\.scss$/,
+          loader: 'css!autoprefixer!sass?outputStyle=compressed'
+        }]
+      }
     },
 
     webpackMiddleware: {
