+++ config/webpack/test.js
@@ -1,0 +1,13 @@
+var webpack = require('webpack');
+var webpackConfigMerger = require('webpack-config-merger');
+var ExtractTextPlugin = require("extract-text-webpack-plugin");
+
+module.exports = webpackConfigMerger(require('./config'), {
+  debug: true,
+  output: {
+    filename: 'sir-trevor.test.js'
+  },
+  plugins: [
+    new ExtractTextPlugin("sir-trevor.test.css")
+  ]
+});
